Update utils.py, adding more functions.

+++ main.py
@@ -26,7 +26,7 @@ logging.basicConfig(level=logging.INFO)
 TOKEN = config('BOT_TOKEN', default='YOUR_BOT_TOKEN_HERE')
 APIKEY = config('APIKEY', default='YOUR_APIKEY_HERE')
 OWNER = config('OWNER', default='https://facebook.com/anbuinfosec')
-CLEAN_TERMINAL = config('CLEAN_TERMINAL', default='False')
+ADMIN_ID = config('ADMIN_ID', default='5839119376')
 
 bot = Bot(token=TOKEN)
 dp = Dispatcher(bot)
@@ -41,7 +41,7 @@ async def on_startup(dp):
 
 @dp.message_handler(commands=['start'])
 async def start(message: types.Message):
-    await message.answer("Hello! Send me video url to download video.")
+    await message.answer("Hello! Send me video url to download files.")
 
 @dp.message_handler(commands=['ping'])
 async def ping(message: types.Message):
@@ -62,40 +62,82 @@ async def ping(message: types.Message):
 async def stats(message: types.Message):
     await message.reply(get_system_info())
 
+@dp.message_handler(commands=['cktmp'])
+async def cktmp(message: types.Message):
+    items = os.listdir('tmp')
+    user_id = str(message.from_user.id)
+    if user_id == ADMIN_ID:
+        processing_message = await message.reply("Checking tmp files!")
+        if len(items) == 0:
+            await processing_message.edit_text('Tmp folder is empty!')
+        else:
+            tmpFiles = "\n".join([f"{index}. {item}" for index, item in enumerate(items, start=1)])
+            await processing_message.edit_text(tmpFiles)
+    else:
+        await message.answer("Hehe boi you are not bot admin.")
+
+@dp.message_handler(commands=['clean'])
+async def clean(message: types.Message):
+    user_id = str(message.from_user.id)
+    if user_id == ADMIN_ID:
+        processing_message = await message.reply("Please wait, deleting files!")
+        tmp_folder = 'tmp'
+        try:
+            for item in os.listdir(tmp_folder):
+                item_path = os.path.join(tmp_folder, item)
+                if os.path.isfile(item_path):
+                    os.remove(item_path)
+                elif os.path.isdir(item_path):
+                    os.rmdir(item_path)
+
+            await processing_message.edit_text("Tmp folder cleaning successful.")
+        except Exception as e:
+            await processing_message.edit_text(f"An error occurred: {e}")
+    else:
+        await message.answer("Hehe boi you are not bot admin.")
 
 @dp.message_handler()
 async def echo(message: types.Message):
     chat_id = message.chat.id
     user_id = message.from_user.id
     message_url = str(message.text)
-    if CLEAN_TERMINAL:
-        clear()
+    keyboard = types.InlineKeyboardMarkup()
+    end_button1 = types.InlineKeyboardButton(text="Author", url=OWNER)
+    end_button2 = types.InlineKeyboardButton(text="Repo", url="https://github.com/anbuinfosec/social-bot-v2")
+    keyboard.add(end_button1, end_button2)
+ 
     await check_tmp()
     downloader = check_downloader(message_url)
     if downloader:
-        processing_message = await message.reply("Please wait, downloading video!")
+        processing_message = await message.reply("Please wait, downloading file!")
         try:
             video_info = await get_video_download_info(message_url, downloader, APIKEY)
             if video_info["status"]:
-                file_path, size = await downloadFromUrl(video_info['url'])
+                file_path, size, file_type = await downloadFromUrl(video_info['url'])
                 if file_path == False:
-                    return await message.reply("Unable to download video!")
-                await processing_message.edit_text("Video download successful. Please wait, uploading your video.")
+                    return await message.reply("Unable to download file!")
+                await processing_message.edit_text("Video download successful. Please wait, uploading your file.")
                 with open(file_path, 'rb') as video_file:
                     await bot.send_chat_action(chat_id, "upload_video")
-                    await bot.send_video(chat_id=message.chat.id, video=InputFile(video_file))
-                    await processing_message.delete()
-                    send_done = True
+                    if file_type == 'video':
+                        await bot.send_video(chat_id=message.chat.id, video=InputFile(video_file))
+                    elif file_type == 'photo':
+                        await bot.send_photo(chat_id=message.chat.id, video=InputFile(video_file))
+                    elif file_type == 'doc':
+                        await bot.send_document(chat_id=message.chat.id, video=InputFile(video_file))
+                    else:
+                        await processing_message.delete()
+                        os.remove(file_path)
+                        return await message.reply("Invalid file type.", reply_markup=keyboard)
                     os.remove(file_path)
-                    keyboard = types.InlineKeyboardMarkup()
-                    url_button = types.InlineKeyboardButton(text="Author", url=OWNER)
-                    url_button2 = types.InlineKeyboardButton(text="Repo", url="https://github.com/anbuinfosec/social-bot-v2")
-                    keyboard.add(url_button, url_button2)
                     await message.reply("Thanks for using our bot.", reply_markup=keyboard)
             else:
+                await processing_message.delete()
+                os.remove(file_path)
                 await message.reply('[-] Download Failed!')
         except Exception as e:
             print (e)
+            os.remove(file_path)
             await message.answer(f"Error: {str(e)}")
     else:
         print("[-] Unsupported URL")
